Add InteractiveRunner Framework
InteractiveRunner enables simple step by step recipes for interacting with
command line interfaces designed for humans. Recipes are composed of Triggers
and Actions, with Triggers being a String match and Actions being some kind of
callback or input action that is invoked as a reaction to the trigger string.

This provides a flexible and easy system to use rclone commands
programmatically that were previously unavailable.