Minor improvements to build scripts.
Most importantly, we defer the build of rclone just before the assemble
stage of the main app, so that Android Studio displays a convenient
clickable error message that allows to install the proper NDK version
directly from the SDK manager.

This commit also de-duplicates rclone build jobs for earch arch.