Fix: RCX not showing as SAF DocumentsProvider in some situations
When the VirtualContentProvider is disabled, it will not show in the file
manager. Unfortunately, this was also the case if the RcdService was not yet
ready, causing file manager hangs in the process. With this commit, the
provider will be shown as empty while loading and initializes in a background
thread, i.e. queryRoots() no longer blocks clients.

(cherry picked from commit c2956fa9d8df0fd70173df3601457da44753e257)