Build rclone automatically when building the main app.
I found it confusing that by default, building the app with
`assembleOssRelease` would result in a non-functional app because rclone
binaries would be missing. This commit ensures sure rclone is built for
each target before building the app, by adding the rclone build task as
a dependency to the "assemble" tasks of the main app.